+++ repository/mocks/category_repository_mock.go
@@ -65,7 +65,7 @@ func (mr *MockCategoryRepositoryMockRecorder) FindAll(ctx interface{}) *gomock.C
 }
 
 // FindById mocks base method.
-func (m *MockCategoryRepository) FindById(ctx context.Context, categoryId int) (domain.Category, error) {
+func (m *MockCategoryRepository) FindById(ctx context.Context, categoryId uint64) (domain.Category, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "FindById", ctx, categoryId)
 	ret0, _ := ret[0].(domain.Category)
